Remove all DM transport payload
