+++ crates/bcr-ebill-api/src/service/notification_service/default_service.rs
@@ -406,8 +406,6 @@ impl NotificationServiceApi for DefaultNotificationService {
             bill_id: bill.id.clone(),
             action_type: Some(ActionType::CheckBill),
             sum: Some(bill.sum),
-            keys: None,
-            blocks: vec![],
         });
         if let Some(node) = self.notification_transport.get(sender_node_id) {
             node.send_private_event(mint, event.try_into()?).await?;
@@ -451,8 +449,6 @@ impl NotificationServiceApi for DefaultNotificationService {
                     bill_id: bill_id.to_owned(),
                     action_type: Some(ActionType::CheckBill),
                     sum,
-                    keys: None,
-                    blocks: vec![],
                 };
                 for (_, recipient) in unique {
                     let event = Event::new_bill(payload.clone());
