+++ crates/bcr-ebill-transport/src/handler/bill_action_event_handler.rs
@@ -1,4 +1,3 @@
-use super::BillChainEventProcessorApi;
 use super::NotificationHandlerApi;
 use crate::BillChainEventPayload;
 use crate::EventType;
@@ -9,6 +8,7 @@ use bcr_ebill_core::ServiceTraitBounds;
 use bcr_ebill_core::notification::BillEventType;
 use bcr_ebill_core::notification::{Notification, NotificationType};
 use bcr_ebill_persistence::NotificationStoreApi;
+use bcr_ebill_persistence::bill::BillChainStoreApi;
 use log::{debug, error, trace, warn};
 use std::sync::Arc;
 
@@ -16,19 +16,19 @@ use std::sync::Arc;
 pub struct BillActionEventHandler {
     notification_store: Arc<dyn NotificationStoreApi>,
     push_service: Arc<dyn PushApi>,
-    processor: Arc<dyn BillChainEventProcessorApi>,
+    chain_store: Arc<dyn BillChainStoreApi>,
 }
 
 impl BillActionEventHandler {
     pub fn new(
         notification_store: Arc<dyn NotificationStoreApi>,
         push_service: Arc<dyn PushApi>,
-        processor: Arc<dyn BillChainEventProcessorApi>,
+        chain_store: Arc<dyn BillChainStoreApi>,
     ) -> Self {
         Self {
             notification_store,
             push_service,
-            processor,
+            chain_store,
         }
     }
 
@@ -42,6 +42,12 @@ impl BillActionEventHandler {
         if event.action_type.is_none() {
             return Ok(());
         }
+
+        // we dont have this chain so skip event
+        if self.chain_store.get_chain(&event.bill_id).await.is_err() {
+            return Ok(());
+        }
+
         // create notification
         let notification = Notification::new_bill_notification(
             &event.bill_id,
@@ -110,20 +116,6 @@ impl NotificationHandlerApi for BillActionEventHandler {
     ) -> Result<()> {
         debug!("incoming bill chain event for {node_id}");
         if let Ok(decoded) = Event::<BillChainEventPayload>::try_from(event.clone()) {
-            if !decoded.data.blocks.is_empty() {
-                if let Err(e) = self
-                    .processor
-                    .process_chain_data(
-                        &decoded.data.bill_id,
-                        decoded.data.blocks.clone(),
-                        decoded.data.keys.clone(),
-                    )
-                    .await
-                {
-                    error!("Failed to process chain data: {}", e);
-                    return Ok(());
-                }
-            }
             if let Err(e) = self.create_notification(&decoded.data, node_id).await {
                 error!("Failed to create notification for bill event: {}", e);
             }
@@ -166,293 +158,54 @@ fn event_description(event_type: &BillEventType) -> String {
 mod tests {
     use std::str::FromStr;
 
-    use bcr_ebill_core::{
-        OptionalPostalAddress, PostalAddress, PublicKey, SecretKey,
-        bill::{BillId, BillKeys, BitcreditBill},
-        blockchain::{
-            Blockchain,
-            bill::{
-                BillBlock, BillBlockchain,
-                block::{
-                    BillEndorseBlockData, BillIssueBlockData, BillParticipantBlockData,
-                    BillRejectBlockData,
-                },
-            },
-        },
-        contact::{BillIdentParticipant, BillParticipant, ContactType},
-        identity::{Identity, IdentityType, IdentityWithAll},
-        notification::ActionType,
-        util::BcrKeys,
-    };
-    use mockall::predicate::{always, eq};
+    use bcr_ebill_core::{PublicKey, bill::BillId, notification::ActionType};
+    use mockall::predicate::eq;
 
-    use crate::handler::{
-        MockBillChainEventProcessorApi,
-        test_utils::{MockNotificationStore, MockPushService, get_test_nostr_event},
+    use crate::handler::test_utils::{
+        MockBillChainStore, MockNotificationStore, MockPushService, get_genesis_chain,
+        get_test_nostr_event,
     };
 
     use super::*;
 
     #[tokio::test]
     async fn test_create_event_handler() {
-        let (notification_store, push_service, bill_chain_event_processor) = create_mocks();
+        let (notification_store, push_service, chain_store) = create_mocks();
         BillActionEventHandler::new(
             Arc::new(notification_store),
             Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
+            Arc::new(chain_store),
         );
     }
 
     #[tokio::test]
-    async fn test_creates_new_chain_for_new_chain_event() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let mut payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        payee.node_id = node_id_test_other();
-        let drawer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, Some(&drawer), None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-        let keys = get_bill_keys();
+    async fn test_fails_to_add_notification_for_unknown_chain() {
+        let (mut notification_store, mut push_service, mut chain_store) = create_mocks();
 
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
+        // no bill chain
+        chain_store
+            .expect_get_chain()
+            .with(eq(bill_id_test()))
+            .returning(|_| {
+                Err(bcr_ebill_persistence::Error::NoSuchEntity(
+                    "id".to_owned(),
+                    "block".to_owned(),
+                ))
+            });
 
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
+        // not look for currently active notification
+        notification_store.expect_get_latest_by_reference().never();
 
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: chain.blocks().clone(),
-                keys: Some(keys.clone()),
-                sum: Some(0),
-                action_type: None,
-            },
-        );
+        // not store new notification
+        notification_store.expect_add().never();
 
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_fails_to_create_new_chain_for_new_chain_event_if_block_validation_fails() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let mut payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        payee.node_id = node_id_test();
-        let drawer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, Some(&drawer), None);
-        let mut chain = get_genesis_chain(Some(bill.clone()));
-        let keys = get_bill_keys();
-
-        // reject to pay without a request to accept will fail
-        let block = BillBlock::create_block_for_reject_to_pay(
-            bill_id_test(),
-            chain.get_latest_block(),
-            &BillRejectBlockData {
-                rejecter: payer.clone().into(),
-                signatory: None,
-                signing_timestamp: chain.get_latest_block().timestamp + 1000,
-                signing_address: empty_address(),
-            },
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            None,
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            chain.get_latest_block().timestamp + 1000,
-        )
-        .unwrap();
-        assert!(chain.try_add_block(block));
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
+        // not send push notification
+        push_service.expect_send().never();
 
         let handler = BillActionEventHandler::new(
             Arc::new(notification_store),
             Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: chain.blocks().clone(),
-                keys: Some(keys.clone()),
-                sum: Some(0),
-                action_type: None,
-            },
-        );
-
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_fails_to_create_new_chain_for_new_chain_event_if_block_signing_check_fails() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        // drawer has a different key than signer, signing check will fail
-        let mut drawer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        drawer.node_id = NodeId::new(BcrKeys::new().pub_key(), bitcoin::Network::Testnet);
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, Some(&drawer), None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-        let keys = get_bill_keys();
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
-
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: chain.blocks().clone(),
-                keys: Some(keys.clone()),
-                sum: Some(0),
-                action_type: None,
-            },
-        );
-
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_adds_block_for_existing_chain_event() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let mut endorsee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        endorsee.node_id = node_id_test_other();
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, None, None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-        let block = BillBlock::create_block_for_endorse(
-            bill_id_test(),
-            chain.get_latest_block(),
-            &BillEndorseBlockData {
-                endorsee: BillParticipantBlockData::Ident(endorsee.clone().into()),
-                // endorsed by payee
-                endorser: BillParticipantBlockData::Ident(
-                    BillIdentParticipant::new(get_baseline_identity().identity)
-                        .unwrap()
-                        .into(),
-                ),
-                signatory: None,
-                signing_timestamp: chain.get_latest_block().timestamp + 1000,
-                signing_address: Some(empty_address()),
-            },
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            None,
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            chain.get_latest_block().timestamp + 1000,
-        )
-        .unwrap();
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
-
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: vec![block.clone()],
-                keys: None,
-                sum: Some(0),
-                action_type: None,
-            },
-        );
-
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_fails_to_add_block_for_invalid_bill_action() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, None, None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-
-        // reject to pay without a request to accept will fail
-        let block = BillBlock::create_block_for_reject_to_pay(
-            bill_id_test(),
-            chain.get_latest_block(),
-            &BillRejectBlockData {
-                rejecter: payer.clone().into(),
-                signatory: None,
-                signing_timestamp: chain.get_latest_block().timestamp + 1000,
-                signing_address: empty_address(),
-            },
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            None,
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            chain.get_latest_block().timestamp + 1000,
-        )
-        .unwrap();
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
-
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
+            Arc::new(chain_store),
         );
 
         let event = Event::new(
@@ -460,137 +213,8 @@ mod tests {
             BillChainEventPayload {
                 bill_id: bill_id_test(),
                 event_type: BillEventType::BillBlock,
-                blocks: vec![block.clone()],
-                keys: None,
                 sum: Some(0),
-                action_type: None,
-            },
-        );
-
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_fails_to_add_block_for_invalidly_signed_blocks() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let endorsee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        // endorser is different than block signer - signature won't be able to be validated
-        let mut endorser = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        endorser.node_id = NodeId::new(BcrKeys::new().pub_key(), bitcoin::Network::Testnet);
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, None, None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-
-        let block = BillBlock::create_block_for_endorse(
-            bill_id_test(),
-            chain.get_latest_block(),
-            &BillEndorseBlockData {
-                endorsee: BillParticipantBlockData::Ident(endorsee.clone().into()),
-                // endorsed by payee
-                endorser: BillParticipantBlockData::Ident(endorser.clone().into()),
-                signatory: None,
-                signing_timestamp: chain.get_latest_block().timestamp + 1000,
-                signing_address: Some(empty_address()),
-            },
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            None,
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            chain.get_latest_block().timestamp + 1000,
-        )
-        .unwrap();
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
-
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: vec![block.clone()],
-                keys: None,
-                sum: Some(0),
-                action_type: None,
-            },
-        );
-
-        handler
-            .handle_event(
-                event.try_into().expect("Envelope from event"),
-                &node_id_test(),
-                Box::new(get_test_nostr_event()),
-            )
-            .await
-            .expect("Event should be handled");
-    }
-
-    #[tokio::test]
-    async fn test_fails_to_add_block_for_unknown_chain() {
-        let payer = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let payee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let endorsee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        let bill = get_test_bitcredit_bill(&bill_id_test(), &payer, &payee, None, None);
-        let chain = get_genesis_chain(Some(bill.clone()));
-
-        let block = BillBlock::create_block_for_endorse(
-            bill_id_test(),
-            chain.get_latest_block(),
-            &BillEndorseBlockData {
-                endorsee: BillParticipantBlockData::Ident(endorsee.clone().into()),
-                // endorsed by payee
-                endorser: BillParticipantBlockData::Ident(
-                    BillIdentParticipant::new(get_baseline_identity().identity)
-                        .unwrap()
-                        .into(),
-                ),
-                signatory: None,
-                signing_timestamp: chain.get_latest_block().timestamp + 1000,
-                signing_address: Some(empty_address()),
-            },
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            None,
-            &BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            chain.get_latest_block().timestamp + 1000,
-        )
-        .unwrap();
-
-        let (notification_store, push_service, mut bill_chain_event_processor) = create_mocks();
-
-        bill_chain_event_processor
-            .expect_process_chain_data()
-            .with(eq(bill_id_test()), always(), always())
-            .returning(|_, _, _| Ok(()));
-
-        let handler = BillActionEventHandler::new(
-            Arc::new(notification_store),
-            Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
-        );
-        let event = Event::new(
-            EventType::Bill,
-            BillChainEventPayload {
-                bill_id: bill_id_test(),
-                event_type: BillEventType::BillBlock,
-                blocks: vec![block.clone()],
-                keys: None,
-                sum: Some(0),
-                action_type: None,
+                action_type: Some(ActionType::CheckBill),
             },
         );
 
@@ -606,7 +230,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_creates_no_notification_for_non_action_event() {
-        let (mut notification_store, mut push_service, bill_chain_event_processor) = create_mocks();
+        let (mut notification_store, mut push_service, chain_store) = create_mocks();
 
         // look for currently active notification
         notification_store.expect_get_latest_by_reference().never();
@@ -620,15 +244,13 @@ mod tests {
         let handler = BillActionEventHandler::new(
             Arc::new(notification_store),
             Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
+            Arc::new(chain_store),
         );
         let event = Event::new(
             EventType::Bill,
             BillChainEventPayload {
                 bill_id: bill_id_test(),
                 event_type: BillEventType::BillBlock,
-                blocks: vec![],
-                keys: None,
                 sum: None,
                 action_type: None,
             },
@@ -646,7 +268,13 @@ mod tests {
 
     #[tokio::test]
     async fn test_creates_notification_for_simple_action_event() {
-        let (mut notification_store, mut push_service, bill_chain_event_processor) = create_mocks();
+        let (mut notification_store, mut push_service, mut chain_store) = create_mocks();
+
+        // given bill chain
+        chain_store
+            .expect_get_chain()
+            .with(eq(bill_id_test()))
+            .returning(|_| Ok(get_genesis_chain(None)));
 
         // look for currently active notification
         notification_store
@@ -671,15 +299,13 @@ mod tests {
         let handler = BillActionEventHandler::new(
             Arc::new(notification_store),
             Arc::new(push_service),
-            Arc::new(bill_chain_event_processor),
+            Arc::new(chain_store),
         );
         let event = Event::new(
             EventType::Bill,
             BillChainEventPayload {
                 bill_id: bill_id_test(),
                 event_type: BillEventType::BillSigned,
-                blocks: vec![],
-                keys: None,
                 sum: Some(0),
                 action_type: Some(ActionType::CheckBill),
             },
@@ -694,130 +320,6 @@ mod tests {
             .await
             .expect("Event should be handled");
     }
-    pub fn get_test_bitcredit_bill(
-        id: &BillId,
-        payer: &BillIdentParticipant,
-        payee: &BillIdentParticipant,
-        drawer: Option<&BillIdentParticipant>,
-        endorsee: Option<&BillIdentParticipant>,
-    ) -> BitcreditBill {
-        let mut bill = empty_bitcredit_bill();
-        bill.id = id.to_owned();
-        bill.payee = BillParticipant::Ident(payee.clone());
-        bill.drawee = payer.clone();
-        if let Some(drawer) = drawer {
-            bill.drawer = drawer.clone();
-        }
-        bill.endorsee = endorsee.map(|e| BillParticipant::Ident(e.to_owned()));
-        bill
-    }
-    fn get_genesis_chain(bill: Option<BitcreditBill>) -> BillBlockchain {
-        let bill = bill.unwrap_or(get_baseline_bill(&bill_id_test()));
-        BillBlockchain::new(
-            &BillIssueBlockData::from(bill, None, 1731593928),
-            get_baseline_identity().key_pair,
-            None,
-            BcrKeys::from_private_key(&private_key_test()).unwrap(),
-            1731593928,
-        )
-        .unwrap()
-    }
-    fn get_baseline_bill(bill_id: &BillId) -> BitcreditBill {
-        let mut bill = empty_bitcredit_bill();
-        let keys = BcrKeys::new();
-
-        bill.maturity_date = "2099-10-15".to_string();
-        let mut payee = empty_bill_identified_participant();
-        payee.name = "payee".to_owned();
-        payee.node_id = NodeId::new(keys.pub_key(), bitcoin::Network::Testnet);
-        bill.payee = BillParticipant::Ident(payee);
-        bill.drawee = BillIdentParticipant::new(get_baseline_identity().identity).unwrap();
-        bill.id = bill_id.clone();
-        bill
-    }
-    fn empty_bitcredit_bill() -> BitcreditBill {
-        BitcreditBill {
-            id: bill_id_test(),
-            country_of_issuing: "AT".to_string(),
-            city_of_issuing: "Vienna".to_string(),
-            drawee: empty_bill_identified_participant(),
-            drawer: empty_bill_identified_participant(),
-            payee: BillParticipant::Ident(empty_bill_identified_participant()),
-            endorsee: None,
-            currency: "sat".to_string(),
-            sum: 500,
-            maturity_date: "2099-11-12".to_string(),
-            issue_date: "2099-08-12".to_string(),
-            city_of_payment: "Vienna".to_string(),
-            country_of_payment: "AT".to_string(),
-            language: "DE".to_string(),
-            files: vec![],
-        }
-    }
-
-    pub fn get_bill_keys() -> BillKeys {
-        BillKeys {
-            private_key: private_key_test(),
-            public_key: node_id_test().pub_key(),
-        }
-    }
-
-    fn get_baseline_identity() -> IdentityWithAll {
-        let keys = BcrKeys::from_private_key(&private_key_test()).unwrap();
-        let mut identity = empty_identity();
-        identity.name = "drawer".to_owned();
-        identity.node_id = NodeId::new(keys.pub_key(), bitcoin::Network::Testnet);
-        identity.postal_address.country = Some("AT".to_owned());
-        identity.postal_address.city = Some("Vienna".to_owned());
-        identity.postal_address.address = Some("Hayekweg 5".to_owned());
-        IdentityWithAll {
-            identity,
-            key_pair: keys,
-        }
-    }
-    fn empty_bill_identified_participant() -> BillIdentParticipant {
-        BillIdentParticipant {
-            t: ContactType::Person,
-            node_id: node_id_test(),
-            name: "some name".to_string(),
-            postal_address: empty_address(),
-            email: None,
-            nostr_relays: vec![],
-        }
-    }
-    fn empty_address() -> PostalAddress {
-        PostalAddress {
-            country: "AT".to_string(),
-            city: "Vienna".to_string(),
-            zip: None,
-            address: "Some address".to_string(),
-        }
-    }
-    fn empty_identity() -> Identity {
-        Identity {
-            t: IdentityType::Ident,
-            node_id: node_id_test(),
-            name: "some name".to_string(),
-            email: Some("some@example.com".to_string()),
-            postal_address: empty_optional_address(),
-            date_of_birth: None,
-            country_of_birth: None,
-            city_of_birth: None,
-            identification_number: None,
-            nostr_relays: vec![],
-            profile_picture_file: None,
-            identity_document_file: None,
-        }
-    }
-
-    pub fn empty_optional_address() -> OptionalPostalAddress {
-        OptionalPostalAddress {
-            country: None,
-            city: None,
-            zip: None,
-            address: None,
-        }
-    }
 
     // bitcrt285psGq4Lz4fEQwfM3We5HPznJq8p1YvRaddszFaU5dY
     pub fn bill_id_test() -> BillId {
@@ -830,30 +332,16 @@ mod tests {
         )
     }
 
-    pub fn private_key_test() -> SecretKey {
-        SecretKey::from_str("d1ff7427912d3b81743d3b67ffa1e65df2156d3dab257316cbc8d0f35eeeabe9")
-            .unwrap()
-    }
-
     pub fn node_id_test() -> NodeId {
         NodeId::from_str("bitcrt02295fb5f4eeb2f21e01eaf3a2d9a3be10f39db870d28f02146130317973a40ac0")
             .unwrap()
     }
 
-    pub fn node_id_test_other() -> NodeId {
-        NodeId::from_str("bitcrt03f9f94d1fdc2090d46f3524807e3f58618c36988e69577d70d5d4d1e9e9645a4f")
-            .unwrap()
-    }
-
-    fn create_mocks() -> (
-        MockNotificationStore,
-        MockPushService,
-        MockBillChainEventProcessorApi,
-    ) {
+    fn create_mocks() -> (MockNotificationStore, MockPushService, MockBillChainStore) {
         (
             MockNotificationStore::new(),
             MockPushService::new(),
-            MockBillChainEventProcessorApi::new(),
+            MockBillChainStore::new(),
         )
     }
 }
